requirements.txt
dominate==2.6.0
six==1.16.0
scipy==1.7.0
opencv_python==4.5.3.56
lpips==0.1.3
scikit_image==0.18.2
torchvision==0.2.1
lmdb==1.2.1
numpy==1.19.2
torch==1.0.0
tqdm==4.61.2
Pillow==8.4.0
skimage==0.0
visdom==0.1.8.9
